requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1
feedparser==6.0.11
python-dateutil==2.9.0
aiocache==0.12.2
PyYAML==6.0.2
discord.py==2.3.2
